Extract SQL helper functions in migration tests to separate file
